Cambio de tipos y caballos en carrera
Las monedas deberían ser double por un tema de precisión.
Mantenimiento de la lista de caballos en carrera

+++ Hipodromos/src/bl/Apuesta.java
@@ -2,15 +2,15 @@ package bl;
 
 public class Apuesta {
 
-    private Float monto;
+    private double monto;
     private Jugador jugador;
 
     //<editor-fold defaultstate="collapsed" desc="Get/Set">
-    public Float getMonto() {
+    public double getMonto() {
         return monto;
     }
 
-    public void setMonto(Float monto) {
+    public void setMonto(double monto) {
         this.monto = monto;
     }
 
@@ -24,7 +24,7 @@ public class Apuesta {
     //</editor-fold>
 
     //<editor-fold defaultstate="collapsed" desc="Constructores">
-    public Apuesta(Float monto, Jugador apostador) {
+    public Apuesta(double monto, Jugador apostador) {
         this.monto = monto;
         this.jugador = apostador;
     }
